# Conflicts:
#	Glitch/Game/Game.cpp

+++ Glitch/Game/Levels/Builders/ParsedEntities/TileSprite.h
@@ -7,10 +7,10 @@ class TileSprite
 {
 public:
 
-	TileSprite(std::string _path, int _width, int _height) : path{ _path }, width{ _width }, height{ _height } {};
+	TileSprite(std::string _path, float _width, float _height) : path{ _path }, width{ _width }, height{ _height } {};
 
 	std::string path;
-	const int width;
-	const int height;
+	const float width;
+	const float height;
 };
